#3745# Add plugin install / management tool

+++ pages/manager/PluginManagementHandler.inc.php
@@ -313,11 +313,14 @@ class PluginManagementHandler extends ManagerHandler {
 		$installedPlugin = $versionDao->getCurrentVersion($plugin);
 		
 		if ($installedPlugin) {
-			
 			$pluginType = explode(".", $installedPlugin->getProductType());
 			$pluginDest = Core::getBaseDir() . DIRECTORY_SEPARATOR . 'plugins' . DIRECTORY_SEPARATOR . $pluginType[1] . DIRECTORY_SEPARATOR . $plugin;
+
+			//make sure plugin type is valid and then delete the files
+			if (in_array($pluginType[1], PluginRegistry::getCategories())) {
+				FileManager::rmtree($pluginDest);
+			}
 			
-			FileManager::rmtree($pluginDest);
 			if(FileManager::fileExists($pluginDest, 'dir')) {
 				$templateMgr->assign('error', true);
 				$templateMgr->assign('message', 'manager.plugins.deleteError');
